Remove the original Buffer and Editor implementations.
This is the start of a re-write. It is enough to build and run the addr tests.
