+++ edit/buffer.go
@@ -1,0 +1,39 @@
+package edit
+
+import (
+	"sync"
+
+	"github.com/eaburns/T/runes"
+)
+
+// A Buffer is an editable rune buffer.
+type Buffer struct {
+	lock  sync.RWMutex
+	runes *runes.Buffer
+	eds   []*Editor
+}
+
+// NewBuffer returns a new, empty Buffer.
+func NewBuffer() *Buffer {
+	return newBuffer(runes.NewBuffer(1 << 12))
+}
+
+func newBuffer(rs *runes.Buffer) *Buffer { return &Buffer{runes: rs} }
+
+// Close closes the Buffer.
+// After Close is called, the Buffer is no longer editable.
+func (buf *Buffer) Close() error {
+	buf.lock.Lock()
+	defer buf.lock.Unlock()
+	return buf.runes.Close()
+}
+
+// Size returns the number of runes in the Buffer.
+//
+// This method must be called with the RLock held.
+func (buf *Buffer) size() int64 { return buf.runes.Size() }
+
+// Rune returns the ith rune in the Buffer.
+//
+// This method must be called with the RLock held.
+func (buf *Buffer) rune(i int64) (rune, error) { return buf.runes.Rune(i) }
