+++ edit/editor.go
@@ -1,0 +1,36 @@
+package edit
+
+import "errors"
+
+// An Editor edits a Buffer of runes.
+type Editor struct {
+	buf   *Buffer
+	marks map[rune]addr
+}
+
+// NewEditor returns an Editor that edits the given buffer.
+func NewEditor(buf *Buffer) *Editor {
+	buf.lock.Lock()
+	defer buf.lock.Unlock()
+	ed := &Editor{
+		buf:   buf,
+		marks: make(map[rune]addr),
+	}
+	buf.eds = append(buf.eds, ed)
+	return ed
+}
+
+// Close closes the editor.
+func (ed *Editor) Close() error {
+	ed.buf.lock.Lock()
+	defer ed.buf.lock.Unlock()
+
+	eds := ed.buf.eds
+	for i := range eds {
+		if eds[i] == ed {
+			ed.buf.eds = append(eds[:i], eds[:i+1]...)
+			return nil
+		}
+	}
+	return errors.New("already closed")
+}
